Require custom-item attributes when closing a new character sheet
TASK #248: Closing/approving a NEW character sheet that contains custom
(non-catalog) cyberware or guns now hard-requires the closer to supply the
mechanical attributes (cyberware: CWP + slot; gun: category, weapon type,
fire mode, power level, optional manufacturer), reaching parity with the
standalone custom cyberware/gun request close flow. Catalog items auto-resolve
from the catalog and are never prompted.

Server (sheets.ts):
- Load cyberware + gun catalog maps before the txn.
- buildSheetInventoryRows() resolves catalog items automatically and requires
  params for custom items (returns {error} when missing/invalid). Runs only on
  the fresh-insert materialization path (linked/resubmitted sheets skip seeding).
- closeSheet accepts sheetParams; build error surfaces as HTTP 400 (no partial
  materialization). Gun/cyberware note formats mirror requests.ts.

Server (review.ts): CloseBodySchema gains sheetCyberware[] + sheetGuns[],
passed through to closeSheet (keyed by full-array index).

OpenAPI: ReviewCloseInput gains the two arrays; codegen + api-client-react /
api-zod regenerated.

Frontend: new SheetCloseDialog fetches the sheet + catalogs (only when open and
approved), detects custom items by name-not-in-catalog, pre-fills cyberware
CWP/slot from the sheet, gates submit until required fields are filled, and sends
params. Wired into ReadyToApplyPanel for subjectType==="sheet". Bulk "Apply all"
sends no params and 400s on custom-item sheets by design (counted as failed).

Tests: sheets.pipeline.test.ts idempotency test updated to send custom params;
added custom-missing→400 (cyberware + gun), custom-with-params success, and
catalog-auto-resolve tests. All 14 pass. tsc clean (api-server + portal);
architect review PASS.

Post-review wiring fixes (addressing prior rejection):
- Centralized sheet close: CloseTicketDialog is now a dispatcher routing
  subjectType==="sheet" to SheetCloseDialog (split into GenericCloseTicketDialog
  to avoid conditional hooks). Every sheet-close entry point (queue
  LifecycleActions + ReadyToApplyPanel) now collects custom attributes; the
  generic note-only dialog can no longer be used for sheets. ReadyToApplyPanel
  reverted to call CloseTicketDialog uniformly.
- SheetCloseDialog now renders catalog-matched cyberware/guns as read-only
  "Auto-resolved from catalog" rows (catalog lookup maps mirror server
  resolution: cyberware highest-CWP-wins, guns first-write-wins, cyber slot
  falls back to the sheet's typed slot).
- Distilled the memory topic file + shortened its MEMORY.md index line.

+++ artifacts/api-server/src/routes/review.ts
@@ -483,7 +483,7 @@ const CloseBodySchema = z.object({
     .array(
       z.object({
         index: z.number().int().min(0),
-        cwp: z.number(),
+        cwp: z.number().int(),
         slot: z.string().trim().max(120),
       }),
     )
